% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_config_GWAS.R
\name{configGWAS}
\alias{configGWAS}
\title{Configure GWAS API Connection}
\usage{
configGWAS()
}
\value{
A connection object to the FinnGen sandbox API.
}
\description{
This function configures and establishes a connection to the FinnGen internal GWAS API
using the \code{FinnGenUtilsR} package. It retrieves the sandbox token from the environment
and uses it to authenticate the connection to the sandbox API.
}
