% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Configuration.R
\name{.fct_assertdatabasesConfig}
\alias{.fct_assertdatabasesConfig}
\title{Internal Assert Databases Configuration}
\usage{
.fct_assertdatabasesConfig(databasesConfig)
}
\arguments{
\item{databasesConfig}{A list representing the databases configuration.}
}
\value{
An AssertCollection object containing any assertion errors.
}
\description{
Internal function to assert the provided databases configuration.
}
