% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_OperateCohorts.R
\name{.plot_upset_cohortsOverlap}
\alias{.plot_upset_cohortsOverlap}
\title{Plot UpSet Cohorts Overlap}
\usage{
.plot_upset_cohortsOverlap(cohortOverlap, nameToShortName)
}
\arguments{
\item{cohortOverlap}{A data frame containing the overlap information between cohorts.}

\item{nameToShortName}{A data frame mapping cohort names to their short names.}
}
\value{
A ggplot object representing the UpSet plot of cohort overlaps.
}
\description{
This function generates an UpSet plot to visualize the overlap between different cohorts.
}
