% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Configuration.R
\name{fcr_setUpLogger}
\alias{fcr_setUpLogger}
\title{Set Up Logger}
\usage{
fcr_setUpLogger()
}
\value{
A logger object.
}
\description{
Sets up a logger with console and file appenders for logging.
}
