% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Configuration.R
\name{fct_assertdatabasesConfig}
\alias{fct_assertdatabasesConfig}
\title{Assert Databases Configuration}
\usage{
fct_assertdatabasesConfig(databasesConfig)
}
\arguments{
\item{databasesConfig}{A list representing the databases configuration.}
}
\value{
None. Throws an error if the configuration is invalid.
}
\description{
Asserts the provided databases configuration and reports any assertion errors.
}
