% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Configuration.R
\name{fct_checkdatabasesConfig}
\alias{fct_checkdatabasesConfig}
\title{Check Databases Configuration}
\usage{
fct_checkdatabasesConfig(databasesConfig)
}
\arguments{
\item{databasesConfig}{A list representing the databases configuration.}
}
\value{
TRUE if the configuration is valid, otherwise a list of error messages.
}
\description{
Checks the provided databases configuration and returns TRUE if the configuration is valid.
If the configuration is invalid, it returns the error messages.
}
