% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Configuration.R
\name{fct_stringToFuction}
\alias{fct_stringToFuction}
\title{Convert String to Function}
\usage{
fct_stringToFuction(packageFunctionString)
}
\arguments{
\item{packageFunctionString}{A string representing the package and function in the format "package::function".}
}
\value{
The function object specified by the string.
}
\description{
Converts a string in the format "package::function" to the actual function.
}
