% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AnalysisWrap.R
\name{mod_analysisWrap_server}
\alias{mod_analysisWrap_server}
\title{Analysis Wrap Server}
\usage{
mod_analysisWrap_server(
  id,
  r_databaseConnection,
  mod_analysisSettings_server,
  fct_executeAnalysis,
  analysisName,
  url_visualiseResults
)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{r_databaseConnection}{A reactive database connection object.}

\item{mod_analysisSettings_server}{A server function for the analysis settings module.}

\item{fct_executeAnalysis}{A function to execute the analysis.}

\item{analysisName}{The name of the analysis.}

\item{url_visualiseResults}{The URL to visualize the results.}
}
\value{
A server function for the analysis wrap module.
}
\description{
Analysis Wrap Server
}
