% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AnalysisWrap.R
\name{mod_analysisWrap_ui}
\alias{mod_analysisWrap_ui}
\title{Analysis Wrap UI}
\usage{
mod_analysisWrap_ui(id, mod_analysisSettings_ui)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{mod_analysisSettings_ui}{A UI function for the analysis settings module.}
}
\value{
A UI definition for the analysis wrap module.
}
\description{
Analysis Wrap UI
}
