% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AppVersion.R
\name{mod_appVersion_ui}
\alias{mod_appVersion_ui}
\title{mod_appVersion_ui}
\usage{
mod_appVersion_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
Shiny UI elements for the database selection module.
}
\description{
UI module for selecting and connecting to databases.
}
