% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CohortWorkbench.R
\name{mod_cohortWorkbench_server}
\alias{mod_cohortWorkbench_server}
\title{Cohort Workbench Server Module}
\usage{
mod_cohortWorkbench_server(
  id,
  r_databaseConnection,
  r_workbenchCache = NULL,
  table_editing = TRUE
)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{r_databaseConnection}{A reactiveValues object containing the database connection and handlers.}

\item{r_workbenchCache}{Optional reactiveValues object with cached workbench summary data.}

\item{table_editing}{A logical value indicating whether table editing is enabled. Default is TRUE.}
}
\value{
A server module for the Cohort Workbench.
}
\description{
This module provides the server logic for the Cohort Workbench.
}
