% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_CohortWorkbench.R
\name{mod_cohortWorkbench_ui}
\alias{mod_cohortWorkbench_ui}
\title{Cohort Workbench UI Module}
\usage{
mod_cohortWorkbench_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A UI definition for the Cohort Workbench.
}
\description{
This module provides the UI for the Cohort Workbench.
}
