% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DragAndDrop.R
\name{mod_dragAndDrop_server}
\alias{mod_dragAndDrop_server}
\title{Drag and Drop Server Module}
\usage{
mod_dragAndDrop_server(id, r_databaseConnection)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{r_databaseConnection}{A reactive database connection object.}
}
\value{
A reactive expression that calculates the cohort operation expression.
}
\description{
This module creates the server logic for a drag and drop interface to define cohort operations.
}
