% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_DragAndDrop.R
\name{mod_dragAndDrop_ui}
\alias{mod_dragAndDrop_ui}
\title{Drag and Drop UI Module}
\usage{
mod_dragAndDrop_ui(id, testing = FALSE)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{testing}{A boolean indicating whether to include testing elements in the UI.}
}
\value{
A UI definition for the drag and drop module.
}
\description{
This module creates the UI for a drag and drop interface to define cohort operations.
}
