% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ExportCohorts.R
\name{mod_exportsCohorts_server}
\alias{mod_exportsCohorts_server}
\title{Export Cohorts Server Module}
\usage{
mod_exportsCohorts_server(id, r_databaseConnection)
}
\arguments{
\item{id}{Module ID}

\item{r_databaseConnection}{Database connection object}
}
\value{
Server logic for the Export Cohorts module
}
\description{
Export Cohorts Server Module
}
