% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ExportCohorts.R
\name{mod_exportsCohorts_ui}
\alias{mod_exportsCohorts_ui}
\title{Export Cohorts UI Module}
\usage{
mod_exportsCohorts_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
UI elements for the Export Cohorts module
}
\description{
Export Cohorts UI Module
}
