% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_AppendCohort.R, R/mod_fct_AppendCohort.R
\name{mod_fct_appendCohort_ui}
\alias{mod_fct_appendCohort_ui}
\title{UI Module for Append Cohort}
\usage{
mod_fct_appendCohort_ui()

mod_fct_appendCohort_ui()
}
\value{
A UI component for appending cohorts.

A UI function for the append cohort module.
}
\description{
This function creates the UI components necessary for appending cohorts.

A shiny Module for appending cohorts.
}
