% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ImportCohortsFromAtlas.R
\name{mod_importCohortsFromAtlas_server}
\alias{mod_importCohortsFromAtlas_server}
\title{Import Cohorts from Atlas Server Module}
\usage{
mod_importCohortsFromAtlas_server(
  id,
  r_databaseConnection,
  filterCohortsRegex = "*"
)
}
\arguments{
\item{id}{A unique identifier for the module.}

\item{r_databaseConnection}{A reactive database connection object.}

\item{filterCohortsRegex}{A regex pattern to filter cohorts by name.}
}
\value{
None
}
\description{
This module provides the server logic for importing cohorts from Atlas.
}
