% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ImportCohortsFromAtlas.R
\name{mod_importCohortsFromAtlas_ui}
\alias{mod_importCohortsFromAtlas_ui}
\title{Import Cohorts from Atlas UI Module}
\usage{
mod_importCohortsFromAtlas_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module.}
}
\value{
A UI definition for the module.
}
\description{
This module provides the UI for importing cohorts from Atlas.
}
