% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ImportCohortsFromCohortTable.R
\name{mod_importCohortsFromCohortsTable_server}
\alias{mod_importCohortsFromCohortsTable_server}
\title{Import Cohorts from Cohort Table Server}
\usage{
mod_importCohortsFromCohortsTable_server(
  id,
  r_databaseConnection,
  filterCohortsRegex = ".*",
  filterCohortsRegexRemove = "()",
  filterCohortsName = "Endpoint"
)
}
\arguments{
\item{id}{Module ID}

\item{r_databaseConnection}{Reactive database connection object}

\item{filterCohortsRegex}{Regular expression to filter cohorts}

\item{filterCohortsRegexRemove}{Regular expression to remove cohorts from filter}

\item{filterCohortsName}{Name filter for cohorts}
}
\value{
Server logic for the Import Cohorts from Cohort Table module
}
\description{
Import Cohorts from Cohort Table Server
}
