% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ImportCohortsFromCohortTable.R
\name{mod_importCohortsFromCohortsTable_ui}
\alias{mod_importCohortsFromCohortsTable_ui}
\title{Import Cohorts from Cohort Table UI}
\usage{
mod_importCohortsFromCohortsTable_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI definition for the Import Cohorts from Cohort Table module
}
\description{
Import Cohorts from Cohort Table UI
}
