% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ImportCohortsFromFile.R
\name{mod_importCohortsFromFile_server}
\alias{mod_importCohortsFromFile_server}
\title{Import Cohorts From File Server Module}
\usage{
mod_importCohortsFromFile_server(id, r_databaseConnection)
}
\arguments{
\item{id}{Module ID}

\item{r_databaseConnection}{A reactive database connection object}
}
\value{
A server function for the module.
}
\description{
A shiny Module to handle server-side operations for importing cohorts from a file.
}
