% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ImportCohortsFromFile.R
\name{mod_importCohortsFromFile_ui}
\alias{mod_importCohortsFromFile_ui}
\title{Import Cohorts From File UI Module}
\usage{
mod_importCohortsFromFile_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI definition for the module.
}
\description{
A shiny Module to import cohorts from a file.
}
