% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MatchCohorts.R
\name{mod_matchCohorts_server}
\alias{mod_matchCohorts_server}
\title{matchCohorts Server Function}
\usage{
mod_matchCohorts_server(id, r_databaseConnection)
}
\arguments{
\item{id}{A module id.}

\item{r_databaseConnection}{A reactive database connection object.}
}
\description{
A shiny Module server function.
}
