% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_OperateCohorts.R
\name{mod_operateCohorts_server}
\alias{mod_operateCohorts_server}
\title{Operate Cohorts Server Module}
\usage{
mod_operateCohorts_server(id, r_databaseConnection)
}
\arguments{
\item{id}{A namespace identifier for the module.}

\item{r_databaseConnection}{A reactive database connection object.}
}
\value{
A server module for operating cohorts.
}
\description{
This module provides the server logic for operating cohorts, including defining cohort operations,
visualizing cohort intersections with an UpSet plot, and creating new cohorts.
}
