% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_OperateCohorts.R
\name{mod_operateCohorts_ui}
\alias{mod_operateCohorts_ui}
\title{Operate Cohorts UI Module}
\usage{
mod_operateCohorts_ui(id)
}
\arguments{
\item{id}{A namespace identifier for the module.}
}
\value{
A UI definition for the operate cohorts module.
}
\description{
This module provides the UI for operating cohorts, including defining cohort operations,
visualizing cohort intersections with an UpSet plot, and creating new cohorts.
}
