% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SelectDatabases.R
\name{mod_selectDatabases_server}
\alias{mod_selectDatabases_server}
\title{mod_selectDatabases_server}
\usage{
mod_selectDatabases_server(id, databasesConfig, r_databaseConnection)
}
\arguments{
\item{id}{Shiny module ID.}

\item{databasesConfig}{Configuration list for databases.}

\item{r_databaseConnection}{Reactive values for database connection.}
}
\value{
Shiny server logic for the database selection module.
}
\description{
Server module for handling database selection and connection.
}
