% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_SelectDatabases.R
\name{mod_selectDatabases_ui}
\alias{mod_selectDatabases_ui}
\title{mod_selectDatabases_ui}
\usage{
mod_selectDatabases_ui(id)
}
\arguments{
\item{id}{Shiny module ID.}
}
\value{
Shiny UI elements for the database selection module.
}
\description{
UI module for selecting and connecting to databases.
}
