% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ViewResults.R
\name{mod_viewResults_server}
\alias{mod_viewResults_server}
\title{View Results Server Module}
\usage{
mod_viewResults_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A server function for the module.
}
\description{
A shiny Module to handle server-side operations for viewing results from a file.
}
