% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ViewResults.R
\name{mod_viewResults_ui}
\alias{mod_viewResults_ui}
\title{View Results UI Module}
\usage{
mod_viewResults_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
A UI definition for the module.
}
\description{
A shiny Module to view results from a file.
}
