% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_Run.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(pathToDatabasesConfigYalm, pathToAnalysisModulesConfigYalm, pathToAtlasWebApiConfigYalm, ...)
}
\arguments{
\item{...}{arguments to pass to golem_opts.
See \code{?golem::get_golem_options} for more details.}
}
\description{
Run the Shiny Application
}
