% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spiner.R
\name{fct_sweetAlertSpinner}
\alias{fct_sweetAlertSpinner}
\title{fct_sweetAlertSpinner}
\usage{
fct_sweetAlertSpinner(
  message,
  logUrl = "/logs/log.txt",
  updateMiliseconds = 500,
  ...
)
}
\arguments{
\item{message}{message to show in the alert}

\item{...}{additional options for shinyWidgets::show_alert}

\item{wait_time_sec}{if not NULL shows a progress bar that loads in the given second}
}
\description{
A sweet alert with a dna spinner
}
