% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_logs.R
\name{fcr_setUpLogger}
\alias{fcr_setUpLogger}
\title{Set Up Logger}
\usage{
fcr_setUpLogger(sToken = "")
}
\arguments{
\item{sToken}{The session token to be added to log entries}
}
\value{
A logger object.
}
\description{
Sets up a logger with console and file appenders for logging.
}
