% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fct_AppendCohort.R
\name{mod_fct_appendCohort_ui}
\alias{mod_fct_appendCohort_ui}
\title{Append Cohort UI Function}
\usage{
mod_fct_appendCohort_ui()
}
\value{
A UI function for the append cohort module.
}
\description{
A shiny Module for appending cohorts.
}
