% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ImportCohortsFromCohortTable.R
\name{mod_importCohortsFromCohortsTable_server}
\alias{mod_importCohortsFromCohortsTable_server}
\title{Import Cohorts from Cohort Table Server}
\usage{
mod_importCohortsFromCohortsTable_server(id, r_databaseConnection)
}
\arguments{
\item{id}{Module ID}

\item{r_databaseConnection}{Reactive database connection object}
}
\value{
Server logic for the Import Cohorts from Cohort Table module
}
\description{
Import Cohorts from Cohort Table Server
}
