% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_MatchCohorts.R
\name{mod_matchCohorts_server}
\alias{mod_matchCohorts_server}
\title{import_cohort_file Server Functions}
\usage{
mod_matchCohorts_server(id, r_databaseConnection)
}
\description{
import_cohort_file Server Functions
}
