% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_Run.R
\name{run_app}
\alias{run_app}
\title{Run Shiny Application}
\usage{
run_app(databasesConfig, analysisModulesConfig, ...)
}
\arguments{
\item{databasesConfig}{A list containing the database configurations.}

\item{analysisModulesConfig}{A list containing the analysis modules configurations.}

\item{...}{Additional arguments passed to \code{shiny::shinyApp}.}
}
\value{
A Shiny application object.
}
\description{
This function initializes and runs a Shiny application with the specified configurations.
}
